+++ packages/cf-deploy-config-writer/src/mta-config/index.ts
@@ -47,7 +47,7 @@ export async function getMtaConfig(rootPath: string): Promise<MtaConfig | undefi
                 break;
             }
         } catch (error) {
-            await new Promise((resolve) => setTimeout(resolve, 300));
+            await new Promise((resolve) => setTimeout(resolve, 500));
         }
     }
     LoggerHelper.logger?.info(`Read mta.yaml with prefix ${mtaConfig?.prefix}`);
@@ -137,13 +137,16 @@ export function doesCDSBinaryExist(): void {
  * @param options
  */
 export function createCAPMTA(cwd: string, options?: string[]): void {
-    let result = spawnSync(CDSExecutable, [...CDSAddMtaParams, ...(options ?? [])], { cwd });
+    const spawnOpts = process.platform.startsWith('win')
+        ? { windowsVerbatimArguments: true, shell: true, cwd }
+        : { cwd };
+    let result = spawnSync(CDSExecutable, [...CDSAddMtaParams, ...(options ?? [])], spawnOpts);
     if (result?.error) {
         throw new Error(`Something went wrong creating mta.yaml! ${result.error}`);
     }
-    // Ensure the package-lock is created otherwise mta build will fail
     const cmd = process.platform === 'win32' ? `npm.cmd` : 'npm';
-    result = spawnSync(cmd, ['install', '--ignore-engines'], { cwd });
+    // Install latest dev dependencies, if any, added by the CF writer
+    result = spawnSync(cmd, ['install', '--ignore-engines'], spawnOpts);
     if (result?.error) {
         throw new Error(`Something went wrong installing node modules! ${result.error}`);
     }