+++ packages/cf-deploy-config-writer/src/utils.ts
@@ -30,7 +30,7 @@ import {
 } from './constants';
 import type { Editor } from 'mem-fs-editor';
 import { type MTABaseConfig, type CFConfig, type CFBaseConfig, RouterModuleType } from './types';
-import { getMtaId, MtaConfig, addMtaDeployParameters } from './mta-config';
+import { getMtaId, type MtaConfig, addMtaDeployParameters, getMtaConfig } from './mta-config';
 import { apiGetInstanceCredentials } from '@sap/cf-tools';
 import LoggerHelper from './logger-helper';
 import { t } from './i18n';
@@ -259,14 +259,17 @@ async function addStandaloneRouter(cfConfig: CFBaseConfig, mtaInstance: MtaConfi
  * @param fs reference to a mem-fs editor
  */
 export async function addRoutingConfig(config: CFBaseConfig, fs: Editor): Promise<void> {
-    const mtaConfigInstance = await MtaConfig.newInstance(config.mtaPath);
-    if (config.routerType === RouterModuleType.Standard) {
-        await addStandaloneRouter(config, mtaConfigInstance, fs);
-    } else {
-        await mtaConfigInstance.addRoutingModules({ isManagedApp: true, addMissingModules: false });
+    const mtaConfigInstance = await getMtaConfig(config.mtaPath);
+    if (mtaConfigInstance) {
+        if (config.routerType === RouterModuleType.Standard) {
+            await addStandaloneRouter(config, mtaConfigInstance, fs);
+        } else {
+            await mtaConfigInstance.addRoutingModules({ isManagedApp: true, addMissingModules: false });
+        }
+        await addMtaDeployParameters(mtaConfigInstance);
+        await mtaConfigInstance.save();
+        LoggerHelper.logger?.debug(t('debug.capMtaUpdated'));
     }
-    await addMtaDeployParameters(mtaConfigInstance);
-    await mtaConfigInstance.save();
 }
 
 /**
@@ -283,6 +286,8 @@ export function setMtaDefaults(config: CFBaseConfig): void {
 /**
  * Update the root package.json with scripts to deploy the MTA.
  *
+ * Note: The fs editor is not passed to `addPackageDevDependency` since the package.json could be updated by other third party tools.
+ *
  * @param {object} Options Input params
  * @param {string} Options.mtaId - MTA ID to be written to package.json
  * @param {string} Options.rootPath - MTA project path
@@ -295,8 +300,8 @@ export async function updateRootPackage(
     const packageExists = fs.exists(join(rootPath, FileName.Package));
     // Append mta scripts only if mta.yaml is at a different level to the HTML5 app
     if (packageExists) {
-        await addPackageDevDependency(rootPath, Rimraf, RimrafVersion, fs);
-        await addPackageDevDependency(rootPath, MbtPackage, MbtPackageVersion, fs);
+        await addPackageDevDependency(rootPath, Rimraf, RimrafVersion);
+        await addPackageDevDependency(rootPath, MbtPackage, MbtPackageVersion);
         let deployArgs: string[] = [];
         if (fs.exists(join(rootPath, MTAFileExtension))) {
             deployArgs = ['-e', MTAFileExtension];
@@ -306,7 +311,7 @@ export async function updateRootPackage(
             { name: 'build', run: `${MTABuildScript} --mtar archive` },
             { name: 'deploy', run: rootDeployMTAScript(deployArgs) }
         ]) {
-            await updatePackageScript(rootPath, script.name, script.run, fs);
+            await updatePackageScript(rootPath, script.name, script.run);
         }
     }
 }