+++ packages/cf-deploy-config-inquirer/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @sap-ux/cf-deploy-config-inquirer
 
+## 0.2.23
+
+### Patch Changes
+
+-   88d2a71: expose new UI prompting to support app frontend service
+
+## 0.2.22
+
+### Patch Changes
+
+-   Updated dependencies [ced5edf]
+    -   @sap-ux/inquirer-common@0.6.32
+
 ## 0.2.21
 
 ### Patch Changes